chore: move writer to write module
