+++ crates/core/src/operations/write/writer.rs
